Separated Play component from PlayRoute
