+++ src/components/modules/Play/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Play.js";
